+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4857JHAV44942PTVVHAVM9"
+        "01KG4NQ3HPTGPHQ8WE25HZ6CWX"
     )
     r"""The unique identifier of the data source"""