+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4857JE3SZZ3YHQEESK2CFQ"
+        "01KG4NQ3HMCTJ54KJ2FC6WDPTV"
     )
     r"""The unique identifier of the data source"""