+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KG4857JGZSRXYH1PXSRY3Q01"
+        "01KG4NQ3HNZRTKDP56335KKH88"
     )
     r"""The unique identifier of the data source"""